Fix MSVC warnings in crc32c_unittest.cc.
This also modifies the CMake configuration, so relevant build targets
will be compiled with /WX (warnings-as-errors) on MSVC. This should help
us catch similar errors in CI in the future.

+++ src/crc32c_sse42.cc
@@ -236,7 +236,7 @@ uint32_t ExtendSse42(uint32_t crc, const uint8_t* data, size_t size) {
     STEP8(l64, p);
   }
 
-  l = l64;
+  l = static_cast<uint32_t>(l64);
   // Process the last few bytes.
   while (p != e) {
     STEP1;
